refactor: update paths in tests
