+++ packages/platform-sdk-ada/test/mariano.test.ts
@@ -1,5 +1,5 @@
 import "jest-extended";
-import { createService } from "./helpers";
+import { createService } from "./mocking";
 import { ClientService } from "../source/client.service";
 import { PrivateKeyService } from "../source/private-key.service";
 import { PublicKeyService } from "../source/public-key.service";
