+++ packages/platform-sdk-http/source/urls.test.ts
@@ -1,4 +1,4 @@
-import { ensureTrailingSlash } from "./helpers";
+import { ensureTrailingSlash } from "./urls";
 
 test("#ensureTrailingSlash", () => {
 	expect(ensureTrailingSlash("#")).toBe("#/");
