+++ packages/platform-sdk-http/source/urls.ts
@@ -1,0 +1,9 @@
+export const ensureTrailingSlash = (url: string): string => {
+	const lastCharacter = url.substr(-1);
+
+	if (lastCharacter != "/") {
+		url = url + "/";
+	}
+
+	return url;
+};
