+++ packages/platform-sdk-profiles/source/drivers/memory/contacts/contact-address.test.ts
@@ -3,7 +3,7 @@ import "reflect-metadata";
 
 import nock from "nock";
 
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { ContactAddress } from "./contact-address";
 import { Profile } from "../profiles/profile";
