+++ packages/platform-sdk-profiles/source/drivers/memory/plugins/plugin-registry.test.ts
@@ -3,7 +3,7 @@ import "reflect-metadata";
 
 import nock from "nock";
 
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { PluginRegistry } from "./plugin-registry";
 
 const createNetworkMocks = () => {
