+++ packages/platform-sdk-profiles/source/drivers/memory/profiles/profile.dumper.test.ts
@@ -3,7 +3,7 @@ import "reflect-metadata";
 
 import nock from "nock";
 
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { Profile } from "./profile";
 import { IProfile, IProfileRepository } from "../../../contracts";
 import { ProfileDumper } from "./profile.dumper";
