+++ packages/platform-sdk-profiles/source/drivers/memory/profiles/profile.importer.test.ts
@@ -5,7 +5,7 @@ import { Base64 } from "@arkecosystem/platform-sdk-crypto";
 import nock from "nock";
 
 import { identity } from "../../../../test/fixtures/identity";
-import { bootContainer, importByMnemonic } from "../../../../test/helpers";
+import { bootContainer, importByMnemonic } from "../../../../test/mocking";
 import { Profile } from "./profile";
 import { IProfile, IProfileRepository, ProfileSetting } from "../../../contracts";
 import { ProfileImporter } from "./profile.importer";
