+++ packages/platform-sdk-profiles/source/drivers/memory/profiles/profile.validator.test.ts
@@ -4,7 +4,7 @@ import "reflect-metadata";
 import { Base64 } from "@arkecosystem/platform-sdk-crypto";
 import nock from "nock";
 
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { Profile } from "./profile";
 import { IProfile, IProfileRepository, ProfileSetting } from "../../../contracts";
 import { ProfileImporter } from "./profile.importer";
