+++ packages/platform-sdk-profiles/source/drivers/memory/repositories/notification-repository.test.ts
@@ -1,6 +1,6 @@
 import "jest-extended";
 import "reflect-metadata";
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { Profile } from "../profiles/profile";
 
 import { NotificationRepository } from "./notification-repository";
