+++ packages/platform-sdk-profiles/source/drivers/memory/repositories/wallet-repository.test.ts
@@ -5,7 +5,7 @@ import nock from "nock";
 import { v4 as uuidv4 } from "uuid";
 
 import { identity } from "../../../../test/fixtures/identity";
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { IProfile, IReadWriteWallet, IWalletData, ProfileSetting } from "../../../contracts";
 import { Profile } from "../profiles/profile";
 import { Wallet } from "../wallets/wallet";
