+++ packages/platform-sdk-profiles/source/drivers/memory/services/fee-service.test.ts
@@ -3,7 +3,7 @@ import "reflect-metadata";
 
 import nock from "nock";
 
-import { bootContainer } from "../../../../test/helpers";
+import { bootContainer } from "../../../../test/mocking";
 import { FeeService } from "./fee-service";
 
 let subject: FeeService;
