+++ packages/platform-sdk-profiles/source/drivers/memory/wallets/wallet.test.ts
@@ -7,7 +7,7 @@ import nock from "nock";
 import { v4 as uuidv4 } from "uuid";
 
 import { identity } from "../../../../test/fixtures/identity";
-import { bootContainer, importByMnemonic } from "../../../../test/helpers";
+import { bootContainer, importByMnemonic } from "../../../../test/mocking";
 import { container } from "../../../environment/container";
 import { Identifiers } from "../../../environment/container.models";
 import { Wallet } from "./wallet";
