+++ packages/platform-sdk-profiles/source/dto/transaction-mapper.test.ts
@@ -5,7 +5,7 @@ import nock from "nock";
 import { v4 as uuidv4 } from "uuid";
 
 import { identity } from "../../test/fixtures/identity";
-import { bootContainer } from "../../test/helpers";
+import { bootContainer } from "../../test/mocking";
 import { IProfile, IReadWriteWallet, ProfileSetting } from "../contracts";
 import { Profile } from "../drivers/memory/profiles/profile";
 import { Wallet } from "../drivers/memory/wallets/wallet";
