+++ packages/platform-sdk-profiles/source/dto/transaction.test.ts
@@ -7,7 +7,7 @@ import { BigNumber } from "@arkecosystem/platform-sdk-support";
 import nock from "nock";
 
 import { identity } from "../../test/fixtures/identity";
-import { bootContainer, importByMnemonic } from "../../test/helpers";
+import { bootContainer, importByMnemonic } from "../../test/mocking";
 import { IExchangeRateService, IProfile, IReadWriteWallet, ProfileSetting } from "../contracts";
 import { Profile } from "../drivers/memory/profiles/profile";
 import { container } from "../environment/container";
